CLEANUP: Replaced snprintf with memcpy as the scenario is not using the substituted argument.
CLEANUP: Fixed compilation warnings.
BUG: The examples MetaData.cpp now deletes the correct properties object.
OPTIM: While computing the hash code, make sure to check that the returned item has value before accessing it.
EXAMPLE: Added the tests for CPP examples.

+++ test/hash/EngineHashInitTests.cpp
@@ -116,7 +116,7 @@ private:
     void writeTestFiles() {
         void* garbledHeader =
             malloc(sizeof(fiftyoneDegreesDataSetHashHeader));
-        for (int i = 0; i < sizeof(fiftyoneDegreesDataSetHashHeader); i++) {
+        for (size_t i = 0; i < sizeof(fiftyoneDegreesDataSetHashHeader); i++) {
             ((byte*)garbledHeader)[i] = 12;
         }
